Move types from types module to related modules
Remove types module and move query types to query module and expire time
types to utils
